test(core): extract the shared integration harness
Move the helpers the integration files imported from each other into
harness.py, deduplicate replace_once/write_image, and mark
test_konfai_streamed_evaluation.py with the integration marker the
other six files already carry.
