+++ tests/integration/test_konfai_auto_patch_training.py
@@ -23,8 +23,7 @@ import sys
 from pathlib import Path
 
 import pytest
-from test_konfai_core_workflows import _prepare_experiment_dir, _subprocess_env
-from test_konfai_ensemble_tta import _replace_once
+from harness import prepare_experiment_dir, replace_once, subprocess_env
 
 pytestmark = pytest.mark.integration
 
@@ -92,10 +91,10 @@ if __name__ == "__main__":
 
 def test_auto_patch_training_restarts_and_completes(tmp_path: Path) -> None:
     experiment_dir = tmp_path / "experiment"
-    _prepare_experiment_dir(experiment_dir, TRAIN_NAME)
+    prepare_experiment_dir(experiment_dir, TRAIN_NAME)
     base = (experiment_dir / "Config.yml").read_text(encoding="utf-8")
-    auto = _replace_once(base, "patch_size: [1, 16, 16]", "patch_size: [1, 0, 0]")
-    auto = _replace_once(auto, "overlap: None", "overlap: 0")
+    auto = replace_once(base, "patch_size: [1, 16, 16]", "patch_size: [1, 0, 0]")
+    auto = replace_once(auto, "overlap: None", "overlap: 0")
     (experiment_dir / "ConfigAuto.yml").write_text(auto, encoding="utf-8")
 
     runner_path = experiment_dir / "run_auto_patch_training.py"
@@ -103,6 +102,6 @@ def test_auto_patch_training_restarts_and_completes(tmp_path: Path) -> None:
     subprocess.run(
         [sys.executable, str(runner_path)],
         cwd=experiment_dir,
-        env=_subprocess_env(),
+        env=subprocess_env(),
         check=True,
     )
